+++ source/c++/src/cppWraper.cpp
@@ -2,11 +2,11 @@
 
 extern "C"{
 
-    Tracker* _Tracker(uint16_t* points, uint16_t* types, float32* confidances, uint16_t size, uint8_t* frame, uint16_t rows, uint16_t cols){
+    Tracker* _Tracker(int32_t* points, uint16_t* types, float32* confidances, uint16_t size, uint8_t* frame, uint16_t rows, uint16_t cols){
         return new Tracker(points, types, confidances, size, frame, rows, cols);
     }
 
-    void _track(Tracker *t, uint16_t* points, uint16_t* types, float32* confidances, uint16_t size, uint8_t* frame){
+    void _track(Tracker *t, int32_t* points, uint16_t* types, float32* confidances, uint16_t size, uint8_t* frame){
         t -> track(points, types, confidances, size, frame);
     }
