+++ source/c++/src/rect.cpp
@@ -14,12 +14,12 @@ void Rect::draw(cv::Mat& frame, cv::Scalar color){
     cv::rectangle(frame, *this, color, 2);
 }
 
-vector<Rect> pointsToRects(uint16_t *points, uint16_t size){
+vector<Rect> pointsToRects(int32_t *points, uint16_t size){
 	
 	std::vector<Rect> rects;
     rects.reserve(size);
 	for (uint16_t i = 0; i < size; i++){
-        uint16_t* currentPoint = points + i*core::_elementsInPoint;
+        int32_t* currentPoint = points + i*core::_elementsInPoint;
         Point A = Point(currentPoint[0], currentPoint[1]);
         Point B = Point(currentPoint[2], currentPoint[3]);
         rects.emplace_back(A, B);
